+++ CHANGELOG.rst
@@ -7,3 +7,4 @@ Changelog
 - Copy source code from `@XenianChannelBot <https://github.com/Nachtalb/XenianChannelBot>`_ and strip it down
 - Improve alias commands
 - Show actual commands in /commands instead of the commands as code
+- Add channel integration with ``/addchanel``,  ``/removechannel`` and ``/list``